requirements.txt
aiofiles==23.2.1
altair==5.2.0
altgraph==0.17.3
amqp==5.1.1
annotated-types==0.6.0
anyio==3.6.2
appdirs==1.4.4
asciimatics==1.14.0
asttokens==2.4.1
async-timeout==4.0.2
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
certifi==2022.9.24
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
comm==0.1.4
contourpy==1.2.0
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.7
dnspython==2.3.0
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
eventlet==0.33.3
exceptiongroup==1.1.3
executing==2.0.1
fake-useragent==1.1.1
fastapi==0.109.0
feedfinder2==0.0.4
feedparser==6.0.10
ffmpy==0.3.1
filelock==3.12.2
fonttools==4.47.2
fsspec==2023.12.2
future==0.18.2
gradio==4.16.0
gradio_client==0.8.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.20.3
idna==3.4
importlib-metadata==6.0.0
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.16.1
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.2.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.5.0
jupyter_core==5.4.0
kiwisolver==1.4.5
kombu==5.2.4
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
MouseInfo==0.1.3
mss==9.0.1
mysqlclient==2.1.1
nest-asyncio==1.5.8
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.1
openpyxl==3.1.2
orjson==3.9.12
packaging==23.2
paho-mqtt==1.6.1
pandas==2.1.2
parse==1.19.0
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
pika==1.3.1
Pillow==9.3.0
pipenv==2022.11.11
platformdirs==3.10.0
prompt-toolkit==3.0.36
psutil==5.9.6
pure-eval==0.2.2
PyAutoGUI==0.9.54
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
pyee==8.2.2
pyfiglet==0.8.post1
pygame==2.3.0
PyGetWindow==0.0.9
Pygments==2.16.1
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.0
pymongo==4.3.3
PyMsgBox==1.0.9
PyMySQL==1.0.2
pynput==1.7.6
pyparsing==3.1.1
pyperclip==1.8.2
pyppeteer==1.0.2
pyquery==2.0.0
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python-multipart==0.0.6
python-telegram-bot==20.0
PyTMX==3.31
pytweening==1.0.7
pytz==2022.7.1
pywin32==305
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==25.1.1
redis==4.5.1
referencing==0.32.1
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
requests-html==0.10.0
rfc3986==1.5.0
rich==13.7.0
rpds-py==0.17.1
ruff==0.1.14
semantic-version==2.10.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.3
starlette==0.35.1
tabulate==0.9.0
tinysegmenter==0.3
tldextract==3.4.0
tomlkit==0.12.0
toolz==0.12.1
tornado==6.3.3
tqdm==4.64.1
traitlets==5.12.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.14
uvicorn==0.27.0
vine==5.0.0
virtualenv==20.24.2
virtualenv-clone==0.5.7
w3lib==2.1.1
wcwidth==0.2.5
websockets==10.4
zipp==3.15.0
